refactor: extract routing and external job services
