+++ scripts/release_artifact_allowlist.txt
@@ -220,7 +220,9 @@ src/diptrace_mcp/services/__init__.py
 src/diptrace_mcp/services/bom.py
 src/diptrace_mcp/services/context.py
 src/diptrace_mcp/services/documents.py
+src/diptrace_mcp/services/external_jobs.py
 src/diptrace_mcp/services/review.py
+src/diptrace_mcp/services/routing.py
 src/diptrace_mcp/sessions.py
 src/diptrace_mcp/silkscreen.py
 src/diptrace_mcp/spatial.py
